+++ packages/components/src/ProForm/index.tsx
@@ -1,9 +1,10 @@
-import React, { useMemo, Fragment } from 'react';
+import React, { useMemo } from 'react';
 import { Collapse, Card } from 'uiw';
 import FormDom from './formdom';
 import ReadFormDom from './readform';
 import { getFormFields } from './widgets';
 import { ProFormProps } from './type';
+import { StoreCtx } from './hooks/store';
 import './style/form-item.less';
 
 export default function ProForm(props: ProFormProps) {
@@ -13,6 +14,7 @@ export default function ProForm(props: ProFormProps) {
     formType = 'card',
     readOnly = false,
     customWidgetsList = {},
+    form,
   } = props;
 
   // 获取表单配置
@@ -21,36 +23,41 @@ export default function ProForm(props: ProFormProps) {
     [formDatas],
   );
 
-  // 判断表单类型
-  const renderForm = useMemo(() => {
-    const formDomProps = { ...props, formfields };
+  const { submitvalidate, clickRef } = form;
 
-    // 判断是否是详情模式
-    const children = readOnly ? (
-      <ReadFormDom {...props} />
-    ) : (
-      <FormDom {...formDomProps} />
+  const store = useMemo(
+    () => ({
+      clickRef,
+      submitvalidate,
+    }),
+    [form],
+  );
+
+  let children: React.ReactNode;
+  const formDomProps = { ...props, formfields };
+  // 非详情模式下渲染标题
+  const renderTitle = !readOnly ? title : undefined;
+  // 判断是否是详情模式
+  const child = readOnly ? (
+    <ReadFormDom {...props} />
+  ) : (
+    <FormDom {...formDomProps} />
+  );
+
+  // 判断渲染的form类型
+  if (formType === 'card') {
+    children = <Card title={renderTitle}>{child}</Card>;
+  } else if (formType === 'collapse') {
+    children = (
+      <Collapse title={renderTitle} activeKey={['1']}>
+        <Collapse.Panel header={title} key={'1'}>
+          {child}
+        </Collapse.Panel>
+      </Collapse>
     );
-    // 非详情模式下渲染标题
-    const renderTitle = !readOnly ? title : undefined;
+  } else {
+    children = child;
+  }
 
-    if (formType === 'card') return <Card title={renderTitle}>{children}</Card>;
-    if (formType === 'collapse') {
-      return (
-        <Collapse title={renderTitle} activeKey={['1']}>
-          <Collapse.Panel header={title} key={'1'}>
-            {children}
-          </Collapse.Panel>
-        </Collapse>
-      );
-    }
-    return (
-      <Fragment>
-        {renderTitle && <h3>{title}</h3>}
-        {children}
-      </Fragment>
-    );
-  }, [formType, formDatas, title]);
-
-  return renderForm;
+  return <StoreCtx.Provider value={store}>{children}</StoreCtx.Provider>;
 }